fix: solucionado bug onde a soma de dois números positivos retornava um número negativo

+++ Calculadora/Calculadora.cs
@@ -15,15 +15,15 @@ namespace Calculadora
             }
             return operacao;
         }
-        public int soma(Operacoes operacao)
+        public long soma(Operacoes operacao)
         {
             return operacao.valorA + operacao.valorB;
         }
-        public int subtracao(Operacoes operacao)
+        public long subtracao(Operacoes operacao)
         {
             return operacao.valorA - operacao.valorB;
         }
-        public int multiplicacao(Operacoes operacao)
+        public long multiplicacao(Operacoes operacao)
         {
             return operacao.valorA * operacao.valorB;
         }
